Schema refactor

+++ Model/Sale.cs
@@ -10,9 +10,8 @@ public class Sale
 {
 
     private Item _item;
-
-    private double _price;
-    public double Price 
+    private decimal _price;
+    public decimal Price 
     { 
         get { return _price; } 
     }
